utils file initial commit
